Feat/update builder card (#260)
* update using darph design

* update frame, create user for wallet

* update with modularity
